backend/requirements.txt
groq
fastapi
uvicorn
holidays
python-dotenv